requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
beanie==1.26.0
pydantic-settings==2.5.2
pyyaml==6.0.2
python-dotenv==1.0.1
